+++ flask_unchained/bundles/sqlalchemy/services/session_manager.py
@@ -1,4 +1,4 @@
-from flask_unchained import BaseService
+from flask_unchained import Service
 from flask_unchained.di import _ServiceMetaclass
 from sqlalchemy_unchained.session_manager import (SessionManager as _SessionManager,
                                                   _SessionManagerMetaclass)
@@ -8,7 +8,7 @@ class SessionManagerMetaclass(_ServiceMetaclass, _SessionManagerMetaclass):
     pass
 
 
-class SessionManager(_SessionManager, BaseService, metaclass=SessionManagerMetaclass):
+class SessionManager(_SessionManager, Service, metaclass=SessionManagerMetaclass):
     """
     The database session manager service.
     """